requirements-dev.txt
# Development and Testing Dependencies

# Testing
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.0  # Para tests paralelos
factory-boy>=3.3.0   # Para fixtures de tests

# Code Quality
flake8>=6.1.0
black>=23.9.0
isort>=5.12.0
pylint>=2.17.0
mypy>=1.5.0
django-stubs>=4.2.0  # Type hints para Django

# Security
bandit>=1.7.5  # Security linting
safety>=2.3.5  # Check dependencies vulnerabilities

# Coverage
coverage>=7.3.0

# Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0



